feat(post-condition): removed post condition component
